make Printer a final class
